Move logic to the index file
